Fix Win64 compilation after r14670
TBR=jkummerow@chromium.org
BUG=v8:2628

Review URL: https://codereview.chromium.org/15133007

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@14671 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/profile-generator.cc
@@ -918,7 +918,7 @@ void ProfileGenerator::RecordTickSample(const TickSample& sample) {
       List<OffsetRange>* ranges = pc_entry->no_frame_ranges();
       if (ranges) {
         Code* code = Code::cast(HeapObject::FromAddress(start));
-        int pc_offset = sample.pc - code->instruction_start();
+        int pc_offset = static_cast<int>(sample.pc - code->instruction_start());
         for (int i = 0; i < ranges->length(); i++) {
           OffsetRange& range = ranges->at(i);
           if (range.from <= pc_offset && pc_offset < range.to) {
